Allow the editor to be resized with the main window
The editor originally had a fixed size, resizing the main window just
recentered the editor in the space.  Now the editor will be resized to
use the largest space possible maintaining the square aspect ratio. A
minimum size is defined by the number of elements and the element size
defined in the configuration.